move chemController logic to systemboard; add utils.makeBool to feature/setState and add feature/toggleState (for parity)
